Dump former MessageKit repo contents
