+++ .codecov.yml
@@ -1,0 +1,32 @@
+codecov:
+  branch: develop
+
+coverage:
+  precision: 2
+  round: nearest
+  range: "60...100"
+  ignore:
+  - Tests/*
+
+status:
+  project:
+    default:
+      target: auto
+      threshold: 2.0
+      branches:
+      - master
+      - develop
+
+    patch:
+      default:
+        target: auto
+        branches:
+        - master
+        - develop
+
+comment:
+  layout: "header, diff, changes, sunburst, uncovered"
+  branches:
+  - master
+  - develop
+  behavior: default
