+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# CHANGELOG
+
+The changelog for `MessageKit`. Also see the [releases](https://github.com/MessageKit/MessageKit/releases) on GitHub.
+
+--------------------------------------
+
+Upcoming release
+----------------
+
+This release closes the [0.1 milestone](https://github.com/MessageKit/MessageKit/milestone/1?closed=1).
+
+Initial release. :tada:
