--- MessageKit.xcodeproj/xcuserdata/omar.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>MessageKit.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>B011FD5B1F1D48E600CE0D6A</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>B011FD641F1D48E700CE0D6A</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
