--- MessageKit/MessageKit.h
@@ -1,19 +1,0 @@
-//
-//  MessageKit.h
-//  MessageKit
-//
-//  Created by Steven on 7/17/17.
-//  Copyright © 2017 MessageKit. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for MessageKit.
-FOUNDATION_EXPORT double MessageKitVersionNumber;
-
-//! Project version string for MessageKit.
-FOUNDATION_EXPORT const unsigned char MessageKitVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <MessageKit/PublicHeader.h>
-
-
