--- MessageKit/MessagesCollectionView.swift
@@ -1,23 +1,0 @@
-//
-//  MessagesCollectionView.swift
-//  MessageKit
-//
-//  Created by Steven on 7/19/17.
-//  Copyright © 2017 MessageKit. All rights reserved.
-//
-
-import UIKit
-
-// MARK: - MessagesCollectionView
-
-open class MessagesCollectionView: UICollectionView {
-
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
-
-}
