--- MessageKit/MessagesViewController.swift
@@ -1,22 +1,0 @@
-//
-//  MessagesViewController.swift
-//  MessageKit
-//
-//  Created by Steven on 7/19/17.
-//  Copyright © 2017 MessageKit. All rights reserved.
-//
-
-import UIKit
-
-// MARK: - MessagesViewController
-
-open class MessagesViewController: UIViewController {
-    
-    // MARK: - View Life Cycle
-
-    override open func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-}
-
