+++ Sources/MessageKit.h
@@ -1,0 +1,23 @@
+//
+//  Created by Jesse Squires
+//  http://www.jessesquires.com
+//
+//
+//  Documentation
+//  http://messagekit.github.io
+//
+//
+//  GitHub
+//  https://github.com/MessageKit/MessageKit
+//
+//
+//  License
+//  Copyright (c) 2016-present Jesse Squires
+//  Released under an MIT license: http://opensource.org/licenses/MIT
+//
+
+#import <UIKit/UIKit.h>
+
+FOUNDATION_EXPORT double MessageKitVersionNumber;
+
+FOUNDATION_EXPORT const unsigned char MessageKitVersionString[];
