+++ Sources/MessagesViewController.swift
@@ -1,0 +1,28 @@
+//
+//  Created by Jesse Squires
+//  http://www.jessesquires.com
+//
+//
+//  Documentation
+//  http://messagekit.github.io
+//
+//
+//  GitHub
+//  https://github.com/MessageKit/MessageKit
+//
+//
+//  License
+//  Copyright (c) 2016-present Jesse Squires
+//  Released under an MIT license: http://opensource.org/licenses/MIT
+//
+
+import UIKit
+import Foundation
+
+
+public final class MessagesViewController: UICollectionViewController {
+
+    override public func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
