+++ Tests/MessageKitTests.swift
@@ -1,0 +1,38 @@
+//
+//  Created by Jesse Squires
+//  http://www.jessesquires.com
+//
+//
+//  Documentation
+//  http://messagekit.github.io
+//
+//
+//  GitHub
+//  https://github.com/MessageKit/MessageKit
+//
+//
+//  License
+//  Copyright (c) 2016-present Jesse Squires
+//  Released under an MIT license: http://opensource.org/licenses/MIT
+//
+
+import XCTest
+@testable import MessageKit
+
+
+final class MessageKitTests: XCTestCase {
+    
+    override func setUp() {
+        super.setUp()
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+    }
+    
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+    
+}
